move behaviour to mod.rs
